+++ CHANGELOG.md
@@ -19,3 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed an issue where the text in the todo list title was not aligned when there were no tags.
+
+## [1.1.1] - 2025-07-26
+
+### Fixed
+
+- Fixed an issue where the search box on the main interface was not working properly.
+
+### Changed
+
+- Remove the useless search button.